requirements.txt
Flask
beautifulsoup4
requests
lxml
openai
python-dotenv
firebase
firebase-admin